connectome/pyproject.toml
[tool.poetry]
name = "cla"
version = "0.1.0"
description = ""
authors = ["myuan <zhengmy@ion.ac.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
opencv-python-headless = "^4.8.0.76"
numpy = "^1.26.0"
streamlit = "^1.26.0"

# 等待 1.4 版本发布，切换到 pypi 中
joblib = {git = "https://github.com/joblib/joblib", rev = "05caf07"}
pyqt5-qt5 = {version = "5.15.2", platform = "linux"}
pyqtwebengine-qt5 = {version = "5.15.2", platform = "linux"}

ntp-manager = "^1.0.6"
orjson = "^3.9.7"
scipy = "^1.11.2"
aicspylibczi = "^3.1.2"
pandas = "^2.1.1"
pint = "^0.22"
polars = "^0.20.17"
bidict = "^0.22.1"
cla-gui = "^0.1.5"
tqdm = "*"
xlsx2csv = "^0.8.1"
connectome-utils = {path = "/home/myuan/projects/connectome-utils", develop = true}
czi-shader = {path = "/home/myuan/projects/czi-shader", develop = true}
wandb = "*"
pillow = "*"
torch = "^2.0.1"
torchvision = "^0.15.2"
segmentation-models-pytorch = "^0.3.3"
albumentations = "^1.3.1"
stackprinter = "^0.2.10"
tensorboard = "^2.14.1"
tomlkit = "^0.12.1"
optuna = "^3.3.0"
optuna-dashboard = "^0.12.0"
nnunetv2 = "^2.2"
pylibczirw = "^3.5.2"
cztile = "^0.1.2"
kornia = "^0.7.0"
antspyx = "^0.4.2"
duckdb = "^0.9.0"
pynrrd = "^1.0.0"
torchregister = "^0.2.2"
torchio = "^0.19.1"
ipyvolume = "^0.6.3"
jupyterlab = "^4.0.7"
ipython-genutils = "^0.2.0"
nhp-utils = "^0.46.18"
dtw-python = "^1.3.0"
zarr = "^2.16.1"
tensorboardx = "^2.6.2.2"
pyarrow = "^14.0.1"
range-compression = "^0.4.0"
pyvips = "^2.2.2"
leidenalg = "^0.10.2"
concave-hull = "^0.0.7"
xarray = "^2024.2.0"
simpleitk = "^2.3.1"
itk = "^5.3.0"
fastapi-htmx = "^0.4.1"
prisma = "^0.13.1"
python-multipart = "^0.0.9"
pyside6 = "^6.6.3"
correction-para = "^0.1.2"


[[tool.poetry.source]]
name = "nhp-pypi"
url = "https://ntp.mkyr.fun/pypi/"
priority = "primary"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "mirrors"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"
pyinstrument = "^4.5.3"
pip = "^23.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
ignore = ["E501", "E701", "E402", "F401", "F821"]

